async function for register and login
