+++ js/api/baseUrl.mjs
@@ -1,1 +1,1 @@
-export const API_BASE_URL = " https://api.noroff.dev";
+export const API_BASE_URL = "https://api.noroff.dev";
